requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
bcrypt==4.0.1
email-validator==2.2.0
python-dotenv==1.0.0
Flask-Mail==0.9.1

# Development dependencies (optional)
# Uncomment these for development tools:
# flask-debugtoolbar==0.13.1
# pytest==7.4.0
# pytest-flask==1.2.0
# coverage==7.3.0
